Add files via upload
In this project, Feature Engineering is done on the data of flight scheduling. Several types of feature engineering techniques are used in different steps of the project. Such as converting string data into integers, and calculating the travel times by converting string data of time into integers. LabelEncoding is done on object data to perform operation on them.